Refactor the VertxManager to be more API like. This allows for better overriding of how the instance is managed for environments like CDI.
